+++ scribo_ct/examples/example_1/source/quux.c
@@ -11,6 +11,14 @@
 
 void quux()
 {
+    SCRIBOD("Auto-fill fatal, error, warning, log, info, and debug:");
+    SCRIBOF();
+    SCRIBOE();
+    SCRIBOW();
+    SCRIBOL();
+    SCRIBOI();
+    SCRIBOD();
+    
     SCRIBOD("Auto-fill method and trace:");
     SCRIBOM();
     SCRIBOT();
@@ -22,6 +30,14 @@ void doQuux()
     
     SCRIBO(DEBUG, "Auto-fill verbosity:");
     SCRIBO();
+    
+    SCRIBO(DEBUG, "Auto-fill fatal, error, warning, log, info, and debug:");
+    SCRIBO(FATAL);
+    SCRIBO(ERROR);
+    SCRIBO(WARNING);
+    SCRIBO(LOG);
+    SCRIBO(INFO);
+    SCRIBO(DEBUG);
     
     SCRIBO(DEBUG, "Auto-fill method and trace:");
     SCRIBO(METHOD);